+++ lib/core/base/base.dart
@@ -1,1 +1,2 @@
 export 'firebase_service.dart';
+export 'firestore_services.dart';