+++ lib/core/constant/message_constant.dart
@@ -5,6 +5,10 @@ class MessageConstant {
 
   static const passwordLength = "Password must be at least 8 characters.";
 
+  static const heightInvalid = "Height must be greater than 0.";
+
+  static const error = "Hmmm... something went wrong";
+
   //LOADING
   static const loading = "Loading...";
 }