+++ lib/feature/common/data/models/user_model.dart
@@ -2,7 +2,6 @@ import 'package:fit_tracker/lib.dart';
 
 class UserModel {
   final String email;
-  final String password;
   final String? name;
   final String? gender;
   final int? height;
@@ -11,7 +10,6 @@ class UserModel {
 
   const UserModel({
     required this.email,
-    required this.password,
     this.name,
     this.gender,
     this.height,
@@ -21,7 +19,6 @@ class UserModel {
 
   factory UserModel.fromJson(Map<String, dynamic> json) => UserModel(
         email: json["email"] as String,
-        password: json["password"] as String,
         name: json["name"] as String? ?? "-",
         gender: json["gender"] as String? ?? "-",
         dateOfBirth: json['date_of_birth'] as String? ?? "-",
@@ -33,11 +30,19 @@ class UserModel {
 
   Map<String, dynamic> toJson() => {
         "email": email,
-        "password": password,
         "name": name,
         "gender": gender,
         "dateOfBirth": dateOfBirth,
-        "height" : height,
-        "weight_records" :  weightRecords?.map((e) => e.toJson()),
+        "height": height,
+        "weight_records": weightRecords?.map((e) => e.toJson()),
       };
+
+  UserEntity toEntity() => UserEntity(
+        email: email,
+        name: name,
+        gender: gender,
+        dateOfBirth: dateOfBirth,
+        height: height,
+        weightRecords: weightRecords?.map((e) => e.toEntity()).toList(),
+      );
 }