+++ lib/feature/common/presentation/router/page_path.dart
@@ -3,4 +3,6 @@ class PagePath {
   static const signIn = "/sign_in";
   static const signUp = "/sign_up";
   static const home = "/home";
+  static const profile = "/profile";
+  static const weightDisplay = "/weight_display";
 }