+++ lib/feature/common/presentation/widget/widget.dart
@@ -1,5 +1,6 @@
 export 'custom_appbar.dart';
 export 'custom_button.dart';
+export 'custom_date_picker.dart';
 export 'custom_loading_indicator.dart';
 export 'custom_text.dart';
 export 'custom_text_field.dart';