+++ lib/feature/profile/profile.dart
@@ -1,0 +1,3 @@
+export 'data/data.dart';
+export 'params/params.dart';
+export 'presentation/presentation.dart';