+++ lib/feature/weight_display/presentation/pages/weight_display_page.dart
@@ -1,0 +1,129 @@
+import 'package:fit_tracker/feature/profile/presentation/cubit/firestore_cubit.dart';
+import 'package:fit_tracker/lib.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+import '../../../common/presentation/widget/dialog/dialog_confirmation.dart';
+
+class WeightDisplayPage extends StatelessWidget {
+  const WeightDisplayPage({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocProvider(
+      create: (context) => FirestoreCubit()..get(),
+      child: const WeightDisplayPageWrapper(),
+    );
+  }
+}
+
+class WeightDisplayPageWrapper extends StatefulWidget {
+  const WeightDisplayPageWrapper({Key? key}) : super(key: key);
+
+  @override
+  State<WeightDisplayPageWrapper> createState() =>
+      _WeightDisplayPageWrapperState();
+}
+
+class _WeightDisplayPageWrapperState extends State<WeightDisplayPageWrapper> {
+  void _handleOnTapList(DeleteWeightParams params) {
+    BasePopup(context).dialog(
+      child: DialogConfirmation(
+        title: "Delete this one ?",
+        message: "This action cannot be undone",
+        positiveButtonText: "Yes, delete",
+        negativeButtonText: "No, cancel",
+        positiveButtonAction: () async {
+          context.read<FirestoreCubit>().deleteWeight(params);
+          Navigator.pop(context, true);
+        },
+        negativeButtonAction: () {
+          Navigator.pop(context);
+        },
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return ScaffoldWrapper(
+      child: Scaffold(
+        backgroundColor: WrapperColors.white,
+        appBar: AppBarSecondary(
+          "Weight records",
+          onTapBack: () {
+            Navigator.pop(context, true);
+          },
+        ),
+        body: CustomScrollView(
+          slivers: [
+            const SliverGap(
+              height: AppGap.big,
+            ),
+            BlocBuilder<FirestoreCubit, FirestoreState>(
+              builder: (context, state) {
+                if (state is FirestoreLoading) {
+                  return const SliverToBoxAdapter(
+                    child: ScreenSizeLoadingIndicator(),
+                  );
+                } else if (state is FirestoreLoaded) {
+                  if (state.user?.weightRecords == null) {
+                    return SliverToBoxAdapter(
+                      child: Text(
+                        "No data found",
+                        style: AppTextStyle.bold.copyWith(
+                          fontSize: AppFontSize.large,
+                        ),
+                      ),
+                    );
+                  } else {
+                    return SliverToBoxAdapter(
+                      child: ListView.separated(
+                        shrinkWrap: true,
+                        physics: const NeverScrollableScrollPhysics(),
+                        padding: EdgeInsets.zero,
+                        itemCount: state.user?.weightRecords?.length ?? 0,
+                        separatorBuilder: (context, index) => const Divider(
+                          height: 0,
+                          color: AppColors.oxfordBlue,
+                        ),
+                        itemBuilder: (BuildContext context, int index) {
+                          state.user?.weightRecords?.sort(
+                            (a, b) => b.recordedDate.compareTo(a.recordedDate),
+                          );
+                          final record = state.user?.weightRecords?[index];
+                          return WeightDisplayRow(
+                            weightRecord: record,
+                            index: index,
+                            onTapDelete: () {
+                              _handleOnTapList(
+                                DeleteWeightParams(data: record!),
+                              );
+                            },
+                          );
+                        },
+                      ),
+                    );
+                  }
+                } else if (state is FirestoreError) {
+                  return SliverToBoxAdapter(
+                    child: Text(
+                      "No data found",
+                      style: AppTextStyle.bold.copyWith(
+                        fontSize: AppFontSize.large,
+                      ),
+                    ),
+                  );
+                } else {
+                  return const SliverToBoxAdapter();
+                }
+              },
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}